+++ components/browser/engine-gecko/src/main/java/mozilla/components/browser/engine/gecko/preferences/GeckoPreferenceAccessor.kt
@@ -8,6 +8,7 @@ import androidx.annotation.OptIn
 import org.mozilla.geckoview.ExperimentalGeckoViewApi
 import org.mozilla.geckoview.GeckoPreferenceController
 import org.mozilla.geckoview.GeckoPreferenceController.GeckoPreference
+import org.mozilla.geckoview.GeckoPreferenceController.SetGeckoPreference
 import org.mozilla.geckoview.GeckoResult
 
 /**
@@ -29,6 +30,15 @@ interface GeckoPreferenceAccessor {
      */
     @OptIn(ExperimentalGeckoViewApi::class)
     fun getGeckoPref(pref: String): GeckoResult<GeckoPreference<*>?>
+
+    /**
+     * Gets the values of a list of Gecko preferences.
+     *
+     * @param prefs The list of the preference to get.
+     * @return A [GeckoResult] object with a list of preference values.
+     */
+    @OptIn(ExperimentalGeckoViewApi::class)
+    fun getGeckoPrefs(prefs: List<String>): GeckoResult<List<GeckoPreference<*>>>
 
     /**
      * Sets the value of a Gecko preference.
@@ -61,6 +71,15 @@ interface GeckoPreferenceAccessor {
     fun setGeckoPref(pref: String, value: Boolean, branch: Int): GeckoResult<Void>
 
     /**
+     * Sets the value of a list of Gecko preferences.
+     *
+     * @param prefs The list of the preference to set.
+     * @return A [GeckoResult] object with a map to indicate whether the operation was successful.
+     */
+    @OptIn(ExperimentalGeckoViewApi::class)
+    fun setGeckoPrefs(prefs: List<SetGeckoPreference<*>>): GeckoResult<Map<String, Boolean>>
+
+    /**
      * Clears a user branch Gecko preference.
      *
      * @param pref The name of the preference to clear.
@@ -80,6 +99,17 @@ internal class DefaultGeckoPreferenceAccessor : GeckoPreferenceAccessor {
     @OptIn(ExperimentalGeckoViewApi::class)
     override fun getGeckoPref(pref: String): GeckoResult<GeckoPreference<*>?> {
         return GeckoPreferenceController.getGeckoPref(pref)
+    }
+
+    /**
+     * Gets the values of a list of Gecko preferences.
+     *
+     * @param prefs The list of the preference to get.
+     * @return A [GeckoResult] object with a list of preference values.
+     */
+    @OptIn(ExperimentalGeckoViewApi::class)
+    override fun getGeckoPrefs(prefs: List<String>): GeckoResult<List<GeckoPreference<*>>> {
+        return GeckoPreferenceController.getGeckoPrefs(prefs)
     }
 
     /**
@@ -131,6 +161,17 @@ internal class DefaultGeckoPreferenceAccessor : GeckoPreferenceAccessor {
         branch: Int,
     ): GeckoResult<Void> {
         return GeckoPreferenceController.setGeckoPref(pref, value, branch)
+    }
+
+    /**
+     * Sets the values of a list of Gecko preferences.
+     *
+     * @param prefs The list of the preference to set.
+     * @return A [GeckoResult] object with a map to indicate whether the operation was successful.
+     */
+    @OptIn(ExperimentalGeckoViewApi::class)
+    override fun setGeckoPrefs(prefs: List<SetGeckoPreference<*>>): GeckoResult<Map<String, Boolean>> {
+        return GeckoPreferenceController.setGeckoPrefs(prefs)
     }
 
     /**